Correct offsets and block alignment

+++ db/table_cache.cc
@@ -445,7 +445,7 @@ Status TableCache::Get(const ReadOptions& options,
         t->LearnedGet(options, k, get_context, prefix_extractor, skip_filters, file_meta);
       }
       else {
-        printf("Looking in file %llu\n", file_meta.fd.packed_number_and_path_id);
+        printf("Looking in file %ld\n", (long)file_meta.fd.packed_number_and_path_id);
         s = t->Get(options, k, get_context, prefix_extractor, skip_filters);
       }
       get_context->SetReplayLog(nullptr);
